Pushing all statistical project
Using different datasets to demonstrate how to use R for statistical methods. Starting with simple linear regression in hw1 to logistic regression in hw5. The final project was using a public dataset to analyze data and solve real world problem and assess which characteristics are best for model. The purpose was to apply statistical data analytic skills.